Move Lexicographically Smallest Equivalent String to their package
