+++ src/main/java/algorithms/curated170/medium/lexicographicallysmallestequivalentstring/LexicographicallySmallestEquivalentString.java
@@ -1,4 +1,4 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.lexicographicallysmallestequivalentstring;
 
 import java.util.HashMap;
 import java.util.Map;
