+++ src/main/java/algorithms/curated170/medium/lexicographicallysmallestequivalentstring/LexicographicallySmallestEquivalentStringDFS.java
@@ -1,4 +1,4 @@
-package algorithms.curated170.medium;
+package algorithms.curated170.medium.lexicographicallysmallestequivalentstring;
 
 import java.util.ArrayList;
 import java.util.Arrays;
